@@ -13,16 +13,12 @@
       }
     },
     "require-dev": {
-      "phpunit/phpunit": "~4.8"
+      "phpunit/phpunit": "^9.0"
     },
     "require": {
+      "php": ">=7.4",
       "ext-simplexml": "*"
     },
     "minimum-stability": "dev",
-    "prefer-stable": true,
-    "config": {
-      "platform": {
-        "php": "5.6"
-      }
-    }
+    "prefer-stable": true
 }
