@@ -1,6 +1,8 @@
 <?php
 
-class CultureFeed_Cdb_Data_Address_GeoInformationTest extends PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+class CultureFeed_Cdb_Data_Address_GeoInformationTest extends TestCase
 {
     public function sampleCoordinates()
     {
