@@ -1,13 +1,15 @@
 <?php
 
-class CultureFeed_Cdb_Data_FileTest extends PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+class CultureFeed_Cdb_Data_FileTest extends TestCase
 {
     /**
      * @var CultureFeed_Cdb_Data_File
      */
     protected $file;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->file = new CultureFeed_Cdb_Data_File();
     }
