@@ -1,7 +1,11 @@
 <?php
 
-class CultureFeed_Cdb_Data_PerformerTest extends PHPUnit_Framework_TestCase
+use PHPUnit\Framework\TestCase;
+
+class CultureFeed_Cdb_Data_PerformerTest extends TestCase
 {
+    use CultureFeed_Cdb_DOMElementAssertionsTrait;
+
     /**
      * @param $fileName
      *
@@ -57,7 +61,7 @@ class CultureFeed_Cdb_Data_PerformerTest extends PHPUnit_Framework_TestCase
         $items = $xpath->query('/performers/performer');
         $this->assertEquals(1, $items->length);
 
-        $this->assertEqualXMLStructure(
+        $this->assertEqualDOMElement(
             $sample->documentElement,
             $items->item(0)
         );
@@ -79,7 +83,7 @@ class CultureFeed_Cdb_Data_PerformerTest extends PHPUnit_Framework_TestCase
         $items = $xpath->query('/performers/performer');
         $this->assertEquals(1, $items->length);
 
-        $this->assertEqualXMLStructure(
+        $this->assertEqualDOMElement(
             $sample->documentElement,
             $items->item(0)
         );
